Fixed cpp-file and turned off notifications

+++ solver_lib/solver.cpp
@@ -1,5 +1,5 @@
 #include "solver.h"
-
+#include <cmath>
 #include <stdexcept>
 
 void solve(float a, float b, float c, float& x1, float& x2)
@@ -11,6 +11,6 @@ void solve(float a, float b, float c, float& x1, float& x2)
         throw std::logic_error{"error: discriminant < 0"};
     }
 
-    x1 = (-b - std::sqrtf(d)) / (2 * a);
-    x2 = (-b + std::sqrtf(d)) / (2 * a);
+    x1 = (-b - std::sqrt(d)) / (2 * a);
+    x2 = (-b + std::sqrt(d)) / (2 * a);
 }
